USI_TWI_Slave.(h|c): add callback for slave receive
The callback will be called after a master write session (at Stop
Condition or repeat Start). Unfortunately, USI provides no interrupt
that rises on Stop Condition flag, so we have to stay in the USI
Overflow Interrupt and wait for Stop Condition.

To keep this time short, we only do it if necessary and only for one
SCL cycle. If we miss a Stop condition, the callback is called on next
Start condition.

The rx buffer will be reseted after the callback is called.